Update Assets.php
Selecting *Keep both* for Assets did not have an effect if some of the assets had the same filenames as an Asset existing previously: in this case the `$foundElements` array would be merged into something like this: 

```
['12', '1234']
```

Where 12 is a previously uploaded asset with a similar filename (e.g. `1.jpg`) but with a completely different content, 1234 is a new Asset (with the same "origin" filename - `1.jpg` -, but with a different generated filename e.g. `1_2020-08-11-175300.jpg`). In this case `$foundElements = array_chunk($foundElements, $limit)[0];` would return only the earlier uploaded Asset, which results in an erroneous import.